consolidate shared test logic
